pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "windows-notice-to-ntfy"
version = "0.1.0"
description = "Forward Windows toast notifications to an ntfy topic."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "windows11toast==1.2.6",
    "winrt-Windows.ApplicationModel==3.2.1",
    "winrt-Windows.Data.Xml.Dom==3.2.1",
    "winrt-Windows.Foundation==3.2.1",
    "winrt-Windows.Foundation.Collections==3.2.1",
    "winrt-Windows.UI.Notifications==3.2.1",
    "winrt-Windows.UI.Notifications.Management==3.2.1",
]

[project.scripts]
windows-notice-to-ntfy = "windows_notice_to_ntfy.cli:main"

[tool.setuptools]
packages = ["windows_notice_to_ntfy"]
